Fixing another crash && adding even more tests

+++ nm/sources/handler.c
@@ -30,7 +30,7 @@ bool file_exists(char *file)
     return file_stat.st_size > 0;
 }
 
-char *buffer_file(char *file, int *size)
+char *buffer_file(char *file, size_t *size)
 {
     int fd = open(file, O_RDONLY);
     struct stat file_stat;
